Corrections, configurations, and API handling.
Translation methods added.

Configuration for Latitude, Longitude, Country and API KEY added.

The way in which the Google Maps API is loaded was changed, now a Loader is used that creates a <script> within the DOM to ensure operation without requiring knowledge of CORS.

View created for the widget configuration section within the settings.