Final Version 1.0.0
The use of google.maps.Marker was swapped for google.maps.marker.AdvancedMarkerElement

The problem of multiple markers was solved.

Modified the MapLoader to reuse the google object whenever available instead of reloading the script.

Now creates an ID for the Map object avoiding errors with the API.

And other minor corrections.